# backend/application/src/main/resources/META-INF/openapi.yaml
openapi: 3.1.0
info:
  title: Surplasse API
  version: 0.1.0
  summary: The Surplasse REST API, defined contract-first.
  description: |
    Surplasse lets independent restaurants run their own direct ordering
    channel. This contract is the single source of truth of the API: the
    backend implements Java interfaces generated from it, the frontends
    consume a generated TypeScript client.

    Version 0.1 covers the public catalog, ordering, payment, restaurateur
    identity flows, the authenticated operational-order list and the
    order-intake control. Other Dashboard reads remain drafts (`x-draft`
    blocks, excluded from generation and compatibility checks) until their
    authorization is implemented end to end.

    Conventions: paths are prefixed with `/v1`, money amounts are integers
    in cents with a `Cents` suffix, timestamps are ISO 8601 UTC, and every
    error is an RFC 9457 Problem Details document.
  contact:
    name: Surplasse
    url: https://surplasse.com
servers:
  - url: /
    description: Current API origin, injected by the environment at runtime.
...
paths:
  /v1/establishments/{slug}/public:
    get:
      tags: [catalog]
      operationId: getEstablishmentPublic
      summary: Get the public profile of an establishment
      description: |
        Returns the public information of an establishment resolved by its
        slug (the subdomain of its mini-site). This endpoint is public: it
        backs the header of the mini-site rendered by the Commande frontend.
        `acceptingOrders` lets that frontend keep the menu readable while
        disabling order actions during an operational pause.
        Unknown slugs, as well as establishments that are not publicly
        visible, yield a 404 without distinction.
      parameters:
        - $ref: "#/components/parameters/EstablishmentSlug"
      responses:
        "200":
          description: The public profile of the establishment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstablishmentPublic"
              example:
                id: 7c9e6679-7425-40de-944b-e07fc1f90ae7
                name: Le Cormoran
                slug: le-cormoran
                address: 12 quai des Belges, 13001 Marseille
                acceptingOrders: true
        "400":
          description: The slug is syntactically invalid.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/validation-error
                title: Validation error
                status: 400
                detail: "Path parameter 'slug' does not match the expected pattern."
                instance: /v1/establishments/Le%20Cormoran/public
        "404":
          description: No establishment matches this slug.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/resource-not-found
                title: Resource not found
                status: 404
                detail: No establishment matches slug 'le-cormoran'.
                instance: /v1/establishments/le-cormoran/public
  /v1/establishments/{slug}/menu:
    get:
      tags: [catalog]
      operationId: getPublishedMenu
      summary: Get the published menu of an establishment
      description: |
        Returns the published menu of an establishment resolved by its slug,
        as a complete read model: categories, products, option groups and
        options, in display order (array order is the display order, no
        position field is exposed). Products and options marked unavailable
        are included with `available: false` so the frontend can render them
        greyed out. If the establishment has no published menu, the response
        is a 404.
      parameters:
        - $ref: "#/components/parameters/EstablishmentSlug"
      responses:
        "200":
          description: The published menu, in display order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicMenu"
              example:
                id: 9b2f5c1a-6d3e-4b7f-8a2c-1e9d8c7b6a5f
                name: Carte principale
                currency: EUR
                categories:
                  - id: c1a2b3c4-d5e6-47f8-89ab-cdef01234501
                    name: Entrées
                    products:
                      - id: d0e1f2a3-b4c5-46d7-88e9-f0a1b2c3d401
                        name: Panisses croustillantes
                        description: Frites de pois chiches, aïoli maison.
                        priceCents: 650
                        available: true
                        optionGroups: []
                  - id: c1a2b3c4-d5e6-47f8-89ab-cdef01234502
                    name: Plats
                    products:
                      - id: d0e1f2a3-b4c5-46d7-88e9-f0a1b2c3d405
                        name: Burger du Vieux-Port
                        description: Pain artisanal, bœuf de Provence, tomme de brebis.
                        priceCents: 1600
                        available: true
                        optionGroups:
                          - id: e1f2a3b4-c5d6-47e8-89f0-a1b2c3d4e502
                            name: Cuisson
                            minChoices: 1
                            maxChoices: 1
                            options:
                              - id: f0a1b2c3-d4e5-46f7-88a9-b0c1d2e3f404
                                name: Saignant
                                extraCostCents: 0
                                available: true
                              - id: f0a1b2c3-d4e5-46f7-88a9-b0c1d2e3f405
                                name: À point
                                extraCostCents: 0
                                available: true
                              - id: f0a1b2c3-d4e5-46f7-88a9-b0c1d2e3f406
                                name: Bien cuit
                                extraCostCents: 0
                                available: true
        "400":
          description: The slug is syntactically invalid.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/validation-error
                title: Validation error
                status: 400
                detail: "Path parameter 'slug' does not match the expected pattern."
                instance: /v1/establishments/Le%20Cormoran/menu
        "404":
          description: No establishment matches this slug, or it has no published menu.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/resource-not-found
                title: Resource not found
                status: 404
                detail: No establishment matches slug 'le-cormoran'.
                instance: /v1/establishments/le-cormoran/menu
  /v1/establishments/{establishmentId}/order-intake:
    get:
      tags: [establishment]
      operationId: getOrderIntake
      summary: Get the operational order-intake state
      description: |
        Returns whether the authenticated restaurateur currently accepts new
        table sessions, orders and payment sessions for this establishment.
        This operational switch does not hide the public menu and never cuts
        access to existing orders, their tracking pages or their event streams.

        The restaurateur session cookie is required. Unknown establishments
        and establishments outside the caller's scope yield the same 404.
      security:
        - restaurateurSession: []
      parameters:
        - $ref: "#/components/parameters/EstablishmentId"
      responses:
        "200":
          description: The current operational order-intake state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderIntakeState"
              example:
                establishmentId: 7c9e6679-7425-40de-944b-e07fc1f90ae7
                status: open
                acceptingOrders: true
                updatedAt: "2026-07-20T10:00:00Z"
        "400":
          description: The establishment identifier is syntactically invalid.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/validation-error
                title: Validation error
                status: 400
                detail: "Path parameter 'establishmentId' must be a UUID."
                instance: /v1/establishments/not-a-uuid/order-intake
        "401":
          description: The restaurateur session is missing or expired.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/session-expired
                title: Session expired
                status: 401
                detail: The restaurateur session is expired or missing.
                instance: /v1/establishments/7c9e6679-7425-40de-944b-e07fc1f90ae7/order-intake
        "404":
          description: Unknown establishment, or establishment outside the caller's scope.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/resource-not-found
                title: Resource not found
                status: 404
                detail: The establishment does not exist or is not accessible.
                instance: /v1/establishments/7c9e6679-7425-40de-944b-e07fc1f90ae7/order-intake
    put:
      tags: [establishment]
      operationId: updateOrderIntake
      summary: Open or pause new orders
      description: |
        Idempotently opens or pauses new table sessions, orders and payment
        sessions for one establishment. Pausing is always allowed. Opening is
        allowed only while the establishment is active, has a published menu,
        exposes at least one active table and can create Stripe charges.

        Operations already in flight are not cancelled. In particular, a
        Payment Intent whose client secret was already returned can still
        succeed and must then be served or refunded. Existing orders, their
        tracking pages and their event streams remain available in both states.

        The restaurateur session cookie is required. Unknown establishments
        and establishments outside the caller's scope yield the same 404.
      security:
        - restaurateurSession: []
      parameters:
        - $ref: "#/components/parameters/EstablishmentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderIntakeUpdate"
      responses:
        "200":
          description: The state after the idempotent update.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderIntakeState"
              example:
                establishmentId: 7c9e6679-7425-40de-944b-e07fc1f90ae7
                status: paused
                acceptingOrders: false
                blockedReason: paused
                updatedAt: "2026-07-20T10:05:00Z"
        "400":
          description: The identifier or payload is syntactically invalid.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/validation-error
                title: Validation error
                status: 400
                detail: "status must be one of: open, paused."
                instance: /v1/establishments/7c9e6679-7425-40de-944b-e07fc1f90ae7/order-intake
        "401":
          description: The restaurateur session is missing or expired.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/session-expired
                title: Session expired
                status: 401
                detail: The restaurateur session is expired or missing.
                instance: /v1/establishments/7c9e6679-7425-40de-944b-e07fc1f90ae7/order-intake
        "404":
          description: Unknown establishment, or establishment outside the caller's scope.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/resource-not-found
                title: Resource not found
                status: 404
                detail: The establishment does not exist or is not accessible.
                instance: /v1/establishments/7c9e6679-7425-40de-944b-e07fc1f90ae7/order-intake
        "422":
          description: The establishment is not ready to open order intake.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                establishmentNotActive:
                  summary: The establishment lifecycle is not active
                  value:
                    type: https://surplasse.com/problems/order-intake-establishment-not-active
                    title: Order intake unavailable
                    status: 422
                    detail: The establishment must be active before accepting orders.
                    instance: /v1/establishments/7c9e6679-7425-40de-944b-e07fc1f90ae7/order-intake
                configurationUnavailable:
                  summary: The menu or table configuration is incomplete
                  value:
                    type: https://surplasse.com/problems/order-intake-configuration-unavailable
                    title: Order intake unavailable
                    status: 422
                    detail: A published menu and at least one active table are required before accepting orders.
                    instance: /v1/establishments/7c9e6679-7425-40de-944b-e07fc1f90ae7/order-intake
                paymentsUnavailable:
                  summary: Stripe Connect cannot accept card charges
                  value:
                    type: https://surplasse.com/problems/order-intake-payments-unavailable
                    title: Order intake unavailable
                    status: 422
                    detail: Stripe card charges must be ready before accepting orders.
                    instance: /v1/establishments/7c9e6679-7425-40de-944b-e07fc1f90ae7/order-intake
...
components:
  securitySchemes:
    tableSession:
      type: apiKey
      in: header
      name: X-Table-Session
      description: >
        Opaque anonymous table session token, delivered by
        `createTableSession` when a table QR is scanned. Bound to one
        establishment and one table, valid for the duration of a meal,
        carrying no personal data.
    restaurateurSession:
      type: apiKey
      in: cookie
      name: surplasse_session
      description: >
        Short session JWT set after the magic link exchange. The cookie is
        host-only for api.surplasse.com, HttpOnly, Secure in production,
        SameSite=Lax and scoped to Path=/. Used by Dashboard endpoints and SSE.
    restaurateurRefresh:
      type: apiKey
      in: cookie
      name: surplasse_refresh
      description: >
        Opaque rotating refresh token, hashed server side and valid for 30
        days. The host-only cookie is HttpOnly, Secure in production,
        SameSite=Lax and scoped to Path=/v1/auth/sessions.
  parameters:
    EstablishmentId:
      name: establishmentId
      in: path
      required: true
      description: Identifier of an establishment.
      schema:
        type: string
        format: uuid
      example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
    EstablishmentSlug:
      name: slug
      in: path
      required: true
      description: >
        Slug of the establishment, the label of its subdomain
        (`{slug}.surplasse.com`). Lowercase letters, digits and hyphens.
      schema:
        type: string
        pattern: "^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?$"
      example: le-cormoran
...
  schemas:
    Problem:
      type: object
      description: >
        RFC 9457 Problem Details document, the single error format of the
        API. The `type` URI is stable and identifies the applicative error.
      properties:
        type:
          type: string
          description: Stable URI identifying the applicative error type.
          default: about:blank
        title:
          type: string
          description: Short, human-readable summary of the error type.
        status:
          type: integer
          description: HTTP status code of this occurrence.
          minimum: 100
          maximum: 599
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          description: URI reference of the request that produced the error.
      required: [title, status]
      example:
        type: https://surplasse.com/problems/resource-not-found
        title: Resource not found
        status: 404
        detail: No establishment matches slug 'le-cormoran'.
        instance: /v1/establishments/le-cormoran/menu
...
    EstablishmentPublic:
      type: object
      description: Public profile of an establishment, rendered on its mini-site.
      properties:
        id:
          type: string
          format: uuid
          description: Identifier of the establishment.
        name:
          type: string
          description: Display name of the establishment.
        slug:
          type: string
          description: Slug of the establishment, label of its subdomain.
        address:
          type: string
          description: Postal address, displayed on the mini-site. Absent when not provided.
        acceptingOrders:
          type: boolean
          description: Whether new table sessions, orders and payment sessions are currently accepted.
      required: [id, name, slug, acceptingOrders]
      example:
        id: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        name: Le Cormoran
        slug: le-cormoran
        address: 12 quai des Belges, 13001 Marseille
        acceptingOrders: true
    OrderIntakeStatus:
      type: string
      description: Operational availability of new table sessions, orders and payment sessions.
      enum: [open, paused]
      example: open
    OrderIntakeBlockedReason:
      type: string
      description: Reason why new orders are not effectively accepted. Absent while accepting orders.
      enum: [paused, establishment_not_active, configuration_unavailable, payments_unavailable]
      example: paused
    OrderIntakeState:
      type: object
      additionalProperties: false
      description: Current operational order-intake state of an establishment.
      properties:
        establishmentId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/OrderIntakeStatus"
        acceptingOrders:
          type: boolean
          description: Effective availability after lifecycle and Stripe readiness are applied.
        blockedReason:
          $ref: "#/components/schemas/OrderIntakeBlockedReason"
        updatedAt:
          type: string
          format: date-time
          description: >
            Instant when the configured status last changed, including an
            automatic pause caused by Stripe. Readiness changes that only
            affect acceptingOrders do not move it. Repeated idempotent updates
            keep it unchanged.
      required: [establishmentId, status, acceptingOrders, updatedAt]
      example:
        establishmentId: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        status: open
        acceptingOrders: true
        updatedAt: "2026-07-20T10:00:00Z"
    OrderIntakeUpdate:
      type: object
      additionalProperties: false
      description: Desired operational state of new orders for one establishment.
      properties:
        status:
          $ref: "#/components/schemas/OrderIntakeStatus"
      required: [status]
      example:
        status: paused
    PublicMenu:
      type: object
      description: >
        The published menu of an establishment, as a complete read model in
        display order. Array order is the display order.
      properties:
        id:
          type: string
          format: uuid
          description: Identifier of the menu.
        name:
          type: string
          description: Name of the menu, chosen by the restaurateur.
        currency:
          type: string
          description: ISO 4217 currency code of every amount in the menu.
        categories:
          type: array
          description: Categories of the menu, in display order.
          items:
            $ref: "#/components/schemas/MenuCategory"
      required: [id, name, currency, categories]
      example:
        id: 9b2f5c1a-6d3e-4b7f-8a2c-1e9d8c7b6a5f
        name: Carte principale
        currency: EUR
        categories:
          - id: c1a2b3c4-d5e6-47f8-89ab-cdef01234501
            name: Entrées
            products:
              - id: d0e1f2a3-b4c5-46d7-88e9-f0a1b2c3d401
                name: Panisses croustillantes
                description: Frites de pois chiches, aïoli maison.
                priceCents: 650
                available: true
                optionGroups: []
    MenuCategory:
      type: object
      description: A section of the menu (starters, mains, drinks).
      properties:
        id:
          type: string
          format: uuid
          description: Identifier of the category.
        name:
          type: string
          description: Display name of the category.
        products:
          type: array
          description: Products of the category, in display order.
          items:
            $ref: "#/components/schemas/MenuProduct"
      required: [id, name, products]
      example:
        id: c1a2b3c4-d5e6-47f8-89ab-cdef01234501
        name: Entrées
        products:
          - id: d0e1f2a3-b4c5-46d7-88e9-f0a1b2c3d401
            name: Panisses croustillantes
            description: Frites de pois chiches, aïoli maison.
            priceCents: 650
            available: true
            optionGroups: []
    MenuProduct:
      type: object
      description: >
        A product as displayed on the menu. Unavailable products are included
        with `available: false` so the frontend renders them greyed out.
      properties:
        id:
          type: string
          format: uuid
          description: Identifier of the product.
        name:
          type: string
          description: Display name of the product.
        description:
          type: string
          description: Short description of the product. Absent when not provided.
        priceCents:
          type: integer
          description: Base price of the product, in cents.
          minimum: 0
        available:
          type: boolean
          description: Whether the product can currently be ordered.
        optionGroups:
          type: array
          description: Option groups of the product, in display order. Empty when the product has no options.
          items:
            $ref: "#/components/schemas/MenuOptionGroup"
      required: [id, name, priceCents, available, optionGroups]
      example:
        id: d0e1f2a3-b4c5-46d7-88e9-f0a1b2c3d405
        name: Burger du Vieux-Port
        description: Pain artisanal, bœuf de Provence, tomme de brebis.
        priceCents: 1600
        available: true
        optionGroups: []
    MenuOptionGroup:
      type: object
      description: >
        A set of options of a product (doneness, size, extras) with its
        choice rules. `minChoices` of 1 or more makes the group mandatory.
      properties:
        id:
          type: string
          format: uuid
          description: Identifier of the option group.
        name:
          type: string
          description: Display name of the group.
        minChoices:
          type: integer
          description: Minimum number of options to pick in the group.
          minimum: 0
        maxChoices:
          type: integer
          description: Maximum number of options to pick in the group. Always greater than or equal to `minChoices`.
          minimum: 1
        options:
          type: array
          description: Options of the group, in display order.
          items:
            $ref: "#/components/schemas/MenuOption"
      required: [id, name, minChoices, maxChoices, options]
      example:
        id: e1f2a3b4-c5d6-47e8-89f0-a1b2c3d4e502
        name: Cuisson
        minChoices: 1
        maxChoices: 1
        options:
          - id: f0a1b2c3-d4e5-46f7-88a9-b0c1d2e3f404
            name: Saignant
            extraCostCents: 0
            available: true
    MenuOption:
      type: object
      description: A variant or extra inside an option group.
      properties:
        id:
          type: string
          format: uuid
          description: Identifier of the option.
        name:
          type: string
          description: Display name of the option.
        extraCostCents:
          type: integer
          description: Extra cost of the option, in cents. Zero for free variants.
          minimum: 0
        available:
          type: boolean
          description: Whether the option can currently be picked.
      required: [id, name, extraCostCents, available]
      example:
        id: f0a1b2c3-d4e5-46f7-88a9-b0c1d2e3f404
        name: Saignant
        extraCostCents: 0
        available: true
...